wsp_backtest/pyproject.toml
[project]
name = "wsp-backtest"
version = "0.1.0"
description = "Isolated WSP backtest engine (Module 2)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "pandas>=2.2",
  "numpy>=1.26",
  "pydantic>=2.7",
  "pyarrow>=16.0",
  "matplotlib>=3.8",
  "scipy>=1.13",
  "statsmodels>=0.14",
  "typer>=0.12",
  "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
